Cleanup + add example
